Simplify flow III: no reactive states.
The StorPool OpenStack integration setup is pretty much linear - we run
through all the layers every time.  Some of them will detect that there
is nothing for them to do, others will perform idempotent actions, but
as a whole, the whole enormous and enormously complicated state diagram
is, well, not needed at all.

Of course there will be some reactive hook management left in the charm
itself, but the layers do not need it at all.